+++ lib/formsmith.rb
@@ -2,6 +2,7 @@
 
 require "action_view"
 require "active_support/core_ext/hash/except"
+require "active_support/core_ext/hash/deep_merge"
 require "active_support/core_ext/object/deep_dup"
 require "active_support/core_ext/string/inflections"
 require "view_component"